\name{LDRTools-package}
\alias{LDRTools-package}
\alias{LDRTools}
\docType{package}
\title{
Tools for Linear Dimension Reduction
}
\description{
Linear dimension reduction and the corresponding subspaces can be uniquely defined using orthogonal projection matrices. This package provides tools to compute distances between such subspaces and to compute the average subspace. 
}
\details{
\tabular{ll}{
Package: \tab LDRTools\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2015-09-24\cr
License: \tab GPL (>= 2)\cr
}
The package implements the methods descriped in \cite{Liski, E., Nordhausen, K., Oja, H. and Ruiz-Gazen, A. (201?), Combining Linear Dimension Reduction
Subspaces, to appear in the proceedings of ICORS 2015}.
}
\author{
Eero Liski, Klaus Nordhausen, Hannu Oja and Anne Ruiz-Gazen. 

Maintainer: Klaus Nordhausen, \email{klaus.nordhausen@utu.fi}
}
\references{
\cite{Liski, E., Nordhausen, K., Oja, H. and Ruiz-Gazen, A. (201?), Combining Linear Dimension Reduction
Subspaces, to appear in the proceedings of ICORS 2015.} 

}

\keyword{ package }

