\name{LDRTools-package}
\alias{LDRTools-package}
\alias{LDRTools}
\docType{package}
\title{
Tools for Linear Dimension Reduction
}
\description{
Linear dimension reduction and the corresponding subspaces can be uniquely defined using orthogonal projection matrices. This package provides tools to compute distances between such subspaces and to compute the average subspace. 
}
\details{
\tabular{ll}{
Package: \tab LDRTools\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-08-08\cr
License: \tab GPL (>= 2)\cr
}
The package implements the methods descriped in \cite{Liski, E., Nordhausen, K., Oja, H. and Ruiz-Gazen, A. (2013), Averaging Orthogonal Projectors, preprint available as arXiv:1210.2575 [stat.ME]. }
}
\author{
Eero Liski, Klaus Nordhausen, Hannu Oja and Anne Ruiz-Gazen. 

Maintainer: Eero Liski, \email{liskieero@gmail.com}
}
\references{
\cite{Liski, E., Nordhausen, K., Oja, H. and Ruiz-Gazen, A. (2013), Averaging Orthogonal Projectors, preprint available as arXiv:1210.2575 [stat.ME]. } 

}

\keyword{ package }

