\name{LDRTools-package}
\alias{LDRTools-package}
\alias{LDRTools}
\docType{package}

\title{
\packageTitle{LDRTools}
}
\description{
\packageDescription{LDRTools}
}
\details{
\packageDESCRIPTION{LDRTools}

\packageIndices{LDRTools}
}

\references{
\cite{Liski E., Nordhausen K., Oja H., and Ruiz-Gazen A. (2016), Combining Linear Dimension Reduction Subspaces. In: Agostinelli C., Basu A., Filzmoser P., Mukherjee D. (eds) Recent Advances in Robust Statistics: Theory and Applications. \doi{10.1007/978-81-322-3643-6_7}.}  

}

\author{
\packageAuthor{LDRTools}

Maintainer: \packageMaintainer{LDRTools}
}



\keyword{ package }

