\name{LDRTools-package}
\alias{LDRTools-package}
\alias{LDRTools}
\docType{package}
\title{
Tools for Linear Dimension Reduction
}
\description{
Linear dimension reduction and the corresponding subspaces can be uniquely defined using orthogonal projection matrices. This package provides tools to compute distances between such subspaces and to compute the average subspace. 
}
\details{
\tabular{ll}{
Package: \tab LDRTools\cr
Type: \tab Package\cr
Version: \tab 0.2-1\cr
Date: \tab 2018-03-02\cr
License: \tab GPL (>= 2)\cr
}
The package implements the methods descriped in Liski, E., Nordhausen, K., Oja, H. and Ruiz-Gazen, A. (2016), Combining Linear Dimension Reduction
Subspaces.
}
\author{
Eero Liski, Klaus Nordhausen, Hannu Oja and Anne Ruiz-Gazen. 

Maintainer: Klaus Nordhausen, <klaus.nordhausen@tuwien.ac.at>
}

\references{
\cite{Liski E., Nordhausen K., Oja H., and Ruiz-Gazen A. (2016), Combining Linear Dimension Reduction Subspaces. In: Agostinelli C., Basu A., Filzmoser P., Mukherjee D. (eds) Recent Advances in Robust Statistics: Theory and Applications. \doi{10.1007/978-81-322-3643-6_7}.}  

}

\keyword{ package }

